+++ bin/dialogs/aboutdialog.py
@@ -13,38 +13,88 @@
 #    GNU General Public License for more details.
 #
 #    You should have received a copy of the GNU General Public License
-#    along with this program; if not, write to the Free Software
-#    Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+#    along with Beam; if not, write to the Free Software Foundation,
+#    Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #    or download it from http://www.gnu.org/licenses/gpl.txt
-#
-#
-#    Revision History:
-#
-#    XX/XX/2014 Version 1.0
-#       - Initial release
 #
 # This Python file uses the following encoding: utf-8
 
+import webbrowser
+import wx
+import wx.html
 
-import wx.html, wx.adv
-
-import textwrap
 from bin.beamsettings import beamSettings
 
 ##################################################
 # About DIALOG
 ##################################################
 
-def ShowAboutDialog(self):
-    info = wx.adv.AboutDialogInfo()
-    info.SetIcon(self.favicon)
-    info.SetName(beamSettings.getString("mainframetitle"))
-    info.SetVersion(beamSettings.getString("version"))
-    info.SetDescription(textwrap.fill(beamSettings.getString("aboutdialogdescription"), 70))
-    info.SetLicence(textwrap.fill(beamSettings.getString("aboutdialoglicense"), 70))
-    info.SetCopyright(beamSettings.getString("aboutcopyright"))
-    info.SetWebSite(beamSettings.getString("aboutwebsite"))
-    info.AddDeveloper(beamSettings.getString("aboutdeveloper"))
-    info.AddArtist(beamSettings.getString("aboutartist"))
 
-    wx.adv.AboutBox(info)
+class _HtmlWindow(wx.html.HtmlWindow):
+    def OnLinkClicked(self, link):
+        webbrowser.open(link.GetHref())
+
+
+def ShowAboutDialog(parent):
+    s = beamSettings.getString
+    dlg = wx.Dialog(parent, title="About Beam", size=(520, 480),
+                    style=wx.DEFAULT_DIALOG_STYLE | wx.RESIZE_BORDER)
+
+    html = _HtmlWindow(dlg)
+    html.SetPage(_build_html(s))
+
+    btn = wx.Button(dlg, wx.ID_OK, "Close")
+    btn.SetDefault()
+
+    sizer = wx.BoxSizer(wx.VERTICAL)
+    sizer.Add(html, 1, wx.EXPAND | wx.ALL, 8)
+    sizer.Add(btn, 0, wx.ALIGN_CENTER | wx.BOTTOM, 10)
+    dlg.SetSizer(sizer)
+    dlg.Layout()
+
+    dlg.ShowModal()
+    dlg.Destroy()
+
+
+def _build_html(s):
+    return """\
+<!DOCTYPE html>
+<html>
+<head><meta charset="utf-8"></head>
+<body style="font-family: sans-serif; margin: 12px; font-size: 11pt;">
+
+<h2 style="margin-bottom: 2px;">Beam &nbsp; <small style="color:#666;">v{version}</small></h2>
+<p style="color:#444;">{copyright}</p>
+
+<p>{description}</p>
+
+<h3>Links</h3>
+<ul>
+  <li><a href="{github}">Documentation &amp; setup (GitHub)</a></li>
+  <li><a href="{facebook}">Facebook</a></li>
+  <li><a href="{website}">Old Project website</a></li>
+  <li><a href="{bitbucket}">Legacy wiki on Bitbucket (reference)</a></li>
+</ul>
+
+<h3>Developers</h3>
+<p>{developer}</p>
+
+<h3>License</h3>
+<p style="font-size:9pt; color:#555;">{license}</p>
+
+<h3>Credits</h3>
+<p>{artist}</p>
+
+</body>
+</html>""".format(
+        version=s("version"),
+        copyright=s("aboutcopyright"),
+        description=s("aboutdialogdescription"),
+        github=s("aboutgithub"),
+        facebook=s("aboutfacebook"),
+        website=s("aboutwebsite"),
+        bitbucket=s("aboutbitbucket"),
+        developer=s("aboutdeveloper").replace("\n", "<br>"),
+        license=s("aboutdialoglicense"),
+        artist=s("aboutartist"),
+    )