+++ bin/songclass.py
@@ -227,7 +227,7 @@ class SongObject(object):
                 #
                 # Trim trailing parentheses in title
                 #
-                if currentRule['Type'] == 'Trim () in Title' and currentRule['Active'] == 'yes':
+                if currentRule['Type'] == 'Cut / Trim' and currentRule['Active'] == 'yes':
                     title_field_name = currentRule.get('Field1', '%Title').replace('%', '')
                     trim_symbol = currentRule.get('Field2', '(')
                     trimmed_title = trim_trailing_suffix_from_symbol(getattr(self, title_field_name), trim_symbol)